Revert "Move and fix nodeManager related unit tests"
This reverts commit 865dc1995743a5492e44044d38dce7a4042fe1c1.
