+++ pkg/kubelet/runonce_test.go
@@ -83,11 +83,11 @@ func TestRunOnce(t *testing.T) {
 		rootDirectory:       "/tmp/kubelet",
 		recorder:            &record.FakeRecorder{},
 		cadvisor:            cadvisor,
+		nodeLister:          testNodeLister{},
 		statusManager:       status.NewManager(nil),
 		containerRefManager: kubecontainer.NewRefManager(),
 		readinessManager:    kubecontainer.NewReadinessManager(),
 		podManager:          podManager,
-		nodeManager:         &fakeNodeManager{},
 		os:                  kubecontainer.FakeOS{},
 		volumeManager:       newVolumeManager(),
 		diskSpaceManager:    diskSpaceManager,
